+++ code/model/modelrender.h
@@ -62,7 +62,7 @@ extern int model_interp_get_texture(const texture_info *tinfo, int elapsed_time)
 
 class model_render_params
 {
-	uint Model_flags;
+	uint64_t Model_flags;
 	uint Debug_flags;
 
 	int Objnum;
@@ -107,7 +107,7 @@ class model_render_params
 public:
 	model_render_params();
 
-	void set_flags(uint flags);
+	void set_flags(uint64_t flags);
 	void set_debug_flags(uint flags);
 	void set_object_number(int num);
 	void set_detail_level_lock(int detail_level_lock);
@@ -133,7 +133,7 @@ public:
 	bool is_alpha_mult_set() const;
 	bool uses_thick_outlines() const;
 
-	uint get_model_flags() const;
+	uint64_t get_model_flags() const;
 	uint get_debug_flags() const;
 	int get_object_number() const;
 	int get_detail_level_lock() const;
@@ -307,7 +307,7 @@ void model_render_queue(const model_render_params* render_info, model_draw_list*
 void submodel_render_immediate(const model_render_params* render_info, const polymodel* pm, const polymodel_instance* pmi, int submodel_num, const matrix* orient, const vec3d* pos);
 void submodel_render_queue(const model_render_params* render_info, model_draw_list* scene, const polymodel* pm, const polymodel_instance* pmi, int submodel_num, const matrix* orient, const vec3d* pos);
 void model_render_buffers(model_draw_list* scene, model_material* rendering_material, const model_render_params* interp, const vertex_buffer* buffer, const polymodel* pm, int mn, int detail_level, uint tmap_flags);
-bool model_render_check_detail_box(const vec3d* view_pos, const polymodel* pm, int submodel_num, uint flags);
+bool model_render_check_detail_box(const vec3d* view_pos, const polymodel* pm, int submodel_num, uint64_t flags);
 void model_render_arc(const vec3d* v1, const vec3d* v2, const color* primary, const color* secondary, float arc_width);
 void model_render_insignias(const insignia_draw_data* insignia);
 void model_render_set_wireframe_color(const color* clr);
