+++ code/starfield/starfield.cpp
@@ -153,7 +153,7 @@ int Stars_background_inited = 0;			// if we're inited
 int Nmodel_num = -1;							// model num
 int Nmodel_instance_num = -1;					// model instance num
 matrix Nmodel_orient = IDENTITY_MATRIX;			// model orientation
-int Nmodel_flags = DEFAULT_NMODEL_FLAGS;		// model flags
+uint64_t Nmodel_flags = DEFAULT_NMODEL_FLAGS;		// model flags
 int Nmodel_bitmap = -1;						// model texture
 float Nmodel_alpha = 1.0f;					// model transparency
 
@@ -1676,7 +1676,7 @@ void subspace_render()
 
 	gr_zbuffer_set(GR_ZBUFF_NONE);
 
-	int render_flags = MR_NO_LIGHTING | MR_ALL_XPARENT;
+	uint64_t render_flags = MR_NO_LIGHTING | MR_ALL_XPARENT;
 
 	Interp_subspace = 1;
 	Interp_subspace_offset_u = 1.0f - subspace_offset_u;
@@ -2317,7 +2317,7 @@ void stars_draw_background()
 	model_render_immediate(&render_info, Nmodel_num, Nmodel_instance_num, &Nmodel_orient, &Eye_position, MODEL_RENDER_ALL, false);
 }
 
-void stars_set_background_model(int new_model, int new_bitmap, int flags, float alpha)
+void stars_set_background_model(int new_model, int new_bitmap, uint64_t flags, float alpha)
 {
 	if (gr_screen.mode == GR_STUB) {
 		return;
@@ -2362,7 +2362,7 @@ void stars_set_background_model(int new_model, int new_bitmap, int flags, float 
 }
 
 // call this to set a specific model as the background model
-void stars_set_background_model(const char* model_name, const char* texture_name, int flags, float alpha)
+void stars_set_background_model(const char* model_name, const char* texture_name, uint64_t flags, float alpha)
 {
 	int new_model = -1;
 	int new_bitmap = -1;
